fix: o metodo de mostrar dashboards foi movido para o controlador do usuario ao invès do de compra
