+++ app/Http/Controllers/CompraController.php
@@ -129,88 +129,4 @@ class CompraController extends Controller
     }
 
 
-
-public function relatorio()
-{
-    $usuario = Auth::user();
-    $compras = $usuario->compras()->get();
-
-    // -----------------------------
-    // Gráfico de barras - Gastos por mês
-    // -----------------------------
-    $valoresPorMes = array_fill(1, 12, 0);
-    foreach ($compras as $compra) {
-        $mes = Carbon::parse($compra->data_compra)->month; // número do mês (1-12)
-        $valoresPorMes[$mes] += $compra->valor;
-    }
-
-    $chartBarra = (new LarapexChart)->barChart()
-        ->setTitle('Gastos - 2025')
-        ->setSubtitle('Valores em reais')
-        ->addData('Gastos', array_values($valoresPorMes))
-        ->setXAxis([
-            'Jan', 'Fev', 'Mar', 'Abr', 'Mai', 'Jun',
-            'Jul', 'Ago', 'Set', 'Out', 'Nov', 'Dez'
-        ])
-        ->setColors(['#10B981'])
-        ->setDataLabels(true)
-        ->setHeight(500)
-        ->setGrid();
-
-    // -----------------------------
-    // Gráfico de pizza - Faixa de valor
-    // -----------------------------
-    $contagem = [
-        'Alta (>50)' => 0,
-        'Normal (20-50)' => 0,
-        'Baixa (<20)' => 0,
-    ];
-
-    foreach ($compras as $compra) {
-        if ($compra->valor > 50) {
-            $contagem['Alta (>50)']++;
-        } elseif ($compra->valor >= 20) {
-            $contagem['Normal (20-50)']++;
-        } else {
-            $contagem['Baixa (<20)']++;
-        }
-    }
-
-    $chartPizza = (new LarapexChart)->pieChart()
-        ->setTitle('Distribuição de Compras por Valor')
-        ->setSubtitle('Quantidade de compras por faixa de valor')
-        ->addData(array_values($contagem))
-        ->setLabels(array_keys($contagem))
-        ->setColors(['#EF4444', '#F59E0B', '#10B981']); // Alta-vermelho, Normal-laranja, Baixa-verde
-
-    // -----------------------------
-    // Gráfico de linha - Gastos por dia do mês atual
-    // -----------------------------
-    $diasNoMes = Carbon::now()->daysInMonth; // número de dias no mês atual
-    $valoresPorDia = array_fill(1, $diasNoMes, 0);
-
-    foreach ($compras as $compra) {
-        $data = Carbon::parse($compra->data_compra);
-        if ($data->month == Carbon::now()->month && $data->year == Carbon::now()->year) {
-            $valoresPorDia[$data->day] += $compra->valor;
-        }
-    }
-
-    $chartLinha = (new LarapexChart)->lineChart()
-        ->setTitle('Gastos do Mês Atual')
-        ->setSubtitle('Valores por dia')
-        ->addData('Gastos', array_values($valoresPorDia))
-        ->setXAxis(range(1, $diasNoMes))
-        ->setColors(['#3B82F6'])
-        ->setDataLabels(true)
-        ->setHeight(400)
-        ->setGrid();
-
-    // -----------------------------
-    // Retorna a view com todos os gráficos
-    // -----------------------------
-    return view('usuario.dashboards', compact('chartBarra', 'chartPizza', 'chartLinha'));
-}
-
-
 }
